+++ kplatform-core/src/appleMain/kotlin/dev/karmakrafts/kplatform/AppleRuntimeMemory.kt
@@ -24,6 +24,7 @@ import kotlinx.cinterop.interpretCPointer
 import kotlinx.cinterop.memScoped
 import kotlinx.cinterop.ptr
 import kotlinx.cinterop.value
+import platform.darwin.KERN_SUCCESS
 import platform.darwin.MACH_TASK_BASIC_INFO
 import platform.darwin.MACH_TASK_BASIC_INFO_COUNT
 import platform.darwin.mach_msg_type_number_tVar
@@ -37,7 +38,10 @@ internal object AppleRuntimeMemory : Memory by AppleGlobalMemory {
         get() = memScoped {
             val info = alloc<mach_task_basic_info>()
             val count = alloc<mach_msg_type_number_tVar> { value = MACH_TASK_BASIC_INFO_COUNT.convert() }
-            task_info(mach_task_self_, MACH_TASK_BASIC_INFO.toUInt(), interpretCPointer(info.rawPtr), count.ptr)
+            if (task_info(
+                    mach_task_self_, MACH_TASK_BASIC_INFO.toUInt(), interpretCPointer(info.rawPtr), count.ptr
+                ) != KERN_SUCCESS
+            ) return@memScoped Memory.UNKNOWN
             info.resident_size.toLong()
         }
 }